+++ packages/amplify-appsync-simulator/package.json
@@ -42,7 +42,7 @@
     "js-string-escape": "^1.0.1",
     "jwt-decode": "^2.2.0",
     "libphonenumber-js": "^1.7.31",
-    "lodash": "^4.17.15",
+    "lodash": "^4.17.19",
     "moment": "^2.24.0",
     "moment-jdateformatparser": "^1.2.1",
     "moment-timezone": "0.5.27",
