+++ packages/amplify-category-auth/package.json
@@ -27,7 +27,7 @@
     "enquirer": "^2.3.4",
     "fs-extra": "^8.1.0",
     "inquirer": "^7.0.3",
-    "lodash": "^4.17.15",
+    "lodash": "^4.17.19",
     "mime-types": "^2.1.26",
     "open": "^7.0.0",
     "ora": "^4.0.3",
