+++ packages/amplify-category-function/package.json
@@ -31,7 +31,7 @@
     "inquirer": "^7.0.3",
     "inquirer-datepicker": "^1.1.0",
     "jstreemap": "^1.28.2",
-    "lodash": "^4.17.15",
+    "lodash": "^4.17.19",
     "mime-types": "^2.1.26",
     "promise-sequential": "^1.1.1",
     "uuid": "^3.4.0"
