+++ packages/amplify-category-storage/package.json
@@ -19,7 +19,7 @@
     "amplify-category-function": "2.21.5",
     "fs-extra": "^8.1.0",
     "inquirer": "^7.0.3",
-    "lodash": "^4.17.15",
+    "lodash": "^4.17.19",
     "uuid": "^3.4.0"
   }
 }
