+++ packages/amplify-cli-core/package.json
@@ -20,7 +20,7 @@
     "test": "jest",
     "test-ci": "jest --ci -i",
     "build": "tsc",
-    "start": "tsc -w",
+    "watch": "tsc -w",
     "clean": "rimraf lib tsconfig.tsbuildinfo"
   },
   "dependencies": {
@@ -28,7 +28,7 @@
     "dotenv": "^8.2.0",
     "fs-extra": "^8.1.0",
     "hjson": "^3.2.1",
-    "lodash": "^4.17.15"
+    "lodash": "^4.17.19"
   },
   "devDependencies": {
     "@types/fs-extra": "^8.0.1",
