+++ packages/amplify-cli-core/src/cliEnvironmentProvider.ts
@@ -1,3 +1,3 @@
 export interface CLIEnvironmentProvider {
-  getCurrent(): string;
+  getCurrentEnvName(): string;
 }
