+++ packages/amplify-cli-core/src/feature-flags/featureFlagFileProvider.ts
@@ -1,24 +1,19 @@
 import * as fs from 'fs-extra';
 import * as _ from 'lodash';
 import * as path from 'path';
-import { FeatureFlags, FeatureFlagsEntry } from '.';
+import { FeatureFlagConfiguration, FeatureFlagsEntry } from '.';
 import { CLIEnvironmentProvider, JSONUtilities } from '..';
 import { FeatureFlagValueProvider } from './featureFlagValueProvider';
-import { AmplifyConfigFileName, AmplifyConfigEnvFileNameTemplate } from '../constants';
+import { amplifyConfigFileName, amplifyConfigEnvFileNameTemplate } from '../constants';
 
 export type FeatureFlagFileProviderOptions = {
   projectPath?: string;
 };
 
-const defaultFeatureFlagFileProviderOptions: FeatureFlagFileProviderOptions = {};
-
 export class FeatureFlagFileProvider implements FeatureFlagValueProvider {
-  constructor(
-    private environmentProvider: CLIEnvironmentProvider,
-    private options: FeatureFlagFileProviderOptions = defaultFeatureFlagFileProviderOptions,
-  ) {}
+  constructor(private environmentProvider: CLIEnvironmentProvider, private options: FeatureFlagFileProviderOptions = {}) {}
 
-  public load = async (): Promise<FeatureFlags> => {
+  public load = async (): Promise<FeatureFlagConfiguration> => {
     if (!this.options.projectPath) {
       throw new Error(`'projectPath' option is missing`);
     }
@@ -27,13 +22,13 @@ export class FeatureFlagFileProvider implements FeatureFlagValueProvider {
       throw new Error(`Project path: '${this.options.projectPath}' does not exist.`);
     }
 
-    let result: FeatureFlags = {
+    const result: FeatureFlagConfiguration = {
       project: {},
       environments: {},
     };
 
     // Read project level file exists
-    const projectConfigFileName = path.join(this.options.projectPath, AmplifyConfigFileName);
+    const projectConfigFileName = path.join(this.options.projectPath, amplifyConfigFileName);
     const projectFeatures = await this.loadConfig(projectConfigFileName);
 
     if (projectFeatures) {
@@ -41,9 +36,9 @@ export class FeatureFlagFileProvider implements FeatureFlagValueProvider {
     }
 
     // Read environment level file if we've a valid environment and file exists
-    const envName = this.environmentProvider.getCurrent();
+    const envName = this.environmentProvider.getCurrentEnvName();
     if (envName !== '') {
-      const envConfigFileName = path.join(this.options.projectPath, AmplifyConfigEnvFileNameTemplate.replace('{env}', envName));
+      const envConfigFileName = path.join(this.options.projectPath, amplifyConfigEnvFileNameTemplate(envName));
       const envFeatures = await this.loadConfig(envConfigFileName);
 
       if (envFeatures) {
@@ -54,13 +49,13 @@ export class FeatureFlagFileProvider implements FeatureFlagValueProvider {
     return result;
   };
 
-  private loadConfig = async (fileName: string): Promise<FeatureFlagsEntry> => {
+  private loadConfig = async (fileName: string): Promise<FeatureFlagsEntry | undefined> => {
     const configFileData = await JSONUtilities.readJson<{ features: FeatureFlagsEntry }>(fileName, {
       throwIfNotExist: false,
     });
 
     if (!configFileData || !configFileData.features) {
-      return (undefined as unknown) as FeatureFlagsEntry;
+      return undefined;
     }
 
     const toLower = (result: any, val: any, key: string) => {
