+++ packages/amplify-cli-core/src/feature-flags/index.ts
@@ -1,3 +1,3 @@
 export * from './envVarFormatError';
+export * from './featureFlagTypes';
 export * from './featureFlags';
-export * from './featureFlagProvider';
