+++ packages/amplify-cli/package.json
@@ -27,7 +27,7 @@
     "test-ci": "jest --ci -i",
     "postinstall": "node scripts/post-install.js",
     "build": "tsc",
-    "start": "tsc -w",
+    "watch": "tsc -w",
     "clean": "rimraf ./lib"
   },
   "engines": {
@@ -78,7 +78,7 @@
     "global-prefix": "^3.0.0",
     "ini": "^1.3.5",
     "inquirer": "^7.0.3",
-    "lodash": "^4.17.15",
+    "lodash": "^4.17.19",
     "open": "^7.0.0",
     "ora": "^4.0.3",
     "parse-json": "^5.0.0",
