+++ packages/amplify-cli/src/init-steps/s9-onSuccess.js
@@ -1,6 +1,6 @@
 const fs = require('fs-extra');
 const sequential = require('promise-sequential');
-const { CLIContextEnvironmentProvider, FeatureFlagProvider } = require('amplify-cli-core');
+const { CLIContextEnvironmentProvider, FeatureFlags } = require('amplify-cli-core');
 const { getFrontendPlugins } = require('../extensions/amplify-helpers/get-frontend-plugins');
 const { getProviderPlugins } = require('../extensions/amplify-helpers/get-provider-plugins');
 const gitManager = require('../extensions/amplify-helpers/git-manager');
@@ -44,12 +44,14 @@ async function run(context) {
   if (context.exeInfo.isNewProject) {
     // Initialize feature flags
     const contextEnvironmentProvider = new CLIContextEnvironmentProvider({
-      getEnvInfo: () => context.amplify.getEnvInfo(),
+      getEnvInfo: context.amplify.getEnvInfo,
     });
 
-    await FeatureFlagProvider.initialize(contextEnvironmentProvider, projectPath);
+    if (!FeatureFlags.isInitialized()) {
+      await FeatureFlags.initialize(contextEnvironmentProvider, projectPath);
+    }
 
-    await FeatureFlagProvider.ensureDefaultFeatureFlags(true);
+    await FeatureFlags.ensureDefaultFeatureFlags(true);
   }
 
   context.exeInfo.projectConfig.providers.forEach(provider => {
