+++ packages/amplify-e2e-core/package.json
@@ -25,7 +25,7 @@
     "fs-extra": "^8.1.0",
     "graphql-transformer-core": "6.19.5",
     "jest-environment-node": "^25.1.0",
-    "lodash": "^4.17.15",
+    "lodash": "^4.17.19",
     "node-pty-prebuilt-multiarch": "^0.9.0",
     "retimer": "2.0.0",
     "rimraf": "^3.0.0",
