+++ packages/amplify-nodejs-function-template-provider/package.json
@@ -23,7 +23,7 @@
   "dependencies": {
     "amplify-function-plugin-interface": "1.4.0",
     "graphql-transformer-core": "6.13.1",
-    "lodash": "^4.17.15"
+    "lodash": "^4.17.19"
   },
   "devDependencies": {
     "@types/inquirer": "^6.5.0",
