+++ packages/amplify-provider-awscloudformation/package.json
@@ -45,7 +45,7 @@
     "import-global": "^0.1.0",
     "ini": "^1.3.5",
     "inquirer": "^7.0.3",
-    "lodash": "^4.17.15",
+    "lodash": "^4.17.19",
     "moment": "^2.24.0",
     "open": "^7.0.0",
     "ora": "^4.0.3",
