+++ packages/amplify-util-mock/package.json
@@ -35,7 +35,7 @@
     "chokidar": "^3.3.1",
     "execa": "^4.0.0",
     "fs-extra": "^8.1.0",
-    "lodash": "^4.17.15",
+    "lodash": "^4.17.19",
     "semver": "^7.1.3",
     "which": "^2.0.2"
   },
